move orbthumb into seperate file
